removed pi_camera; created image_processing package; nodes ported: decoder; launchfiles ported: decoder and rectifier.
