+++ packages/image_processing/CMakeLists.txt
@@ -1,18 +1,13 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(pi_camera)
+project(image_processing)
 
 find_package(catkin REQUIRED COMPONENTS
   roscpp
   rospy
-  duckietown_msgs # Every duckietown packages should use this.
-  std_msgs
+  duckietown_msgs
+  sensor_msgs
 )
 
-catkin_python_setup()
-
-
-
 catkin_package()
-
 
 include_directories(${catkin_INCLUDE_DIRS})
