+++ packages/image_processing/config/decoder_node/default.yaml
@@ -1,0 +1,2 @@
+#parameters for image_processing/decoder_node
+publish_freq: 2.0
