+++ packages/image_processing/launch/image_rectification.launch
@@ -4,11 +4,8 @@
 
     <arg name="veh" doc="Name of vehicle. ex: megaman"/>
 
-
     <group ns="$(arg veh)">
-        <node  name="$(arg node_name)" pkg="$(arg pkg_name)" type="$(arg pkg_name)"
-              output="screen">
+        <node pkg="$(arg pkg_name)" type="$(arg pkg_name)" name="$(arg node_name)" output="screen">
         </node>
-
     </group>
 </launch>
