+++ packages/image_processing/package.xml
@@ -1,24 +1,23 @@
 <?xml version="1.0"?>
 <package>
-  <name>pi_camera</name>
+  <name>image_processing</name>
   <version>1.0.0</version>
-  <description>TODO: Add a description of package `pi_camera` in `package.xml`.</description>
+  <description>TODO: Add a description of package `image_processing` in `package.xml`.</description>
 
   <maintainer email="acensi@idsc.mavt.ethz.ch">Andrea Censi</maintainer>
-  
 
   <license>GPLv3</license>
 
   <buildtool_depend>catkin</buildtool_depend>
   <build_depend>duckietown_msgs</build_depend>
+  <build_depend>sensor_msgs</build_depend>
   <build_depend>roscpp</build_depend>
   <build_depend>rospy</build_depend>
-  <build_depend>std_msgs</build_depend>
 
   <run_depend>duckietown_msgs</run_depend>
+  <run_depend>sensor_msgs</run_depend>
   <run_depend>roscpp</run_depend>
   <run_depend>rospy</run_depend>
-  <run_depend>std_msgs</run_depend>
 
   <export>
   </export>
