--- packages/pi_camera/.autogenerated.md
@@ -1,185 +1,0 @@
-<div id='pi_camera-autogenerated' markdown='1'>
-
-
-<!-- do not edit this file, autogenerated -->
-
-## Package information 
-
-[Link to package on Github](github:org=duckietown,repo=Software,path=05-teleop/pi_camera,branch=master18)
-
-### Essentials {nonumber="1"}
-
-Maintainer: [Andrea Censi](mailto:acensi@idsc.mavt.ethz.ch)
-
-### Description {nonumber="1"}
-
-TODO: Add a description of package `pi_camera` in `package.xml`.
-
-
-
-</div>
-
-<!-- file start -->
-
-<div id='pi_camera-camera_node_sequence-autogenerated' markdown='1'>
-
-
-<!-- do not edit this file, autogenerated -->
-
-(Generated from [configuration `camera_node_sequence.easy_node.yaml`](github:org=duckietown,repo=Software,path=camera_node_sequence.easy_node.yaml,branch=master18).)
-
-Camera driver, second approach.
-
-### Parameters {nonumber="1"}
-
-No parameters defined.
-
-### Subscriptions {nonumber="1"}
-
-No subscriptions defined.
-
-### Publishers {nonumber="1"}
-
-No publishers defined.
-
-
-
-</div><!-- file start -->
-
-<div id='pi_camera-cam_info_reader_node-autogenerated' markdown='1'>
-
-
-<!-- do not edit this file, autogenerated -->
-
-(Generated from [configuration `cam_info_reader_node.easy_node.yaml`](github:org=duckietown,repo=Software,path=cam_info_reader_node.easy_node.yaml,branch=master18).)
-
-Publishes a CameraInfo message every time it receives an image.
-
-
-### Parameters {nonumber="1"}
-
-**Parameter `config`**: `str`; default value: `'baseline'`
-
-TODO: Missing description for entry "`config`".
-
-**Parameter `cali_file_name`**: `str`; default value: `'default'`
-
-TODO: Missing description for entry "`cali_file_name`".
-
-**Parameter `image_type`**: `str`; default value: `'compressed'`
-
-TODO: Missing description for entry "`image_type`".
-
-### Subscriptions {nonumber="1"}
-
-**Subscription `compressed_image`**: topic `~compressed_image` (`CompressedImage`)
-
-If image_type is "compressed" then it's CompressedImage, otherwise Image.
-
-### Publishers {nonumber="1"}
-
-**Publisher `camera_info`**: topic `~camera_info` (`CameraInfo`)
-
-TODO: Missing description for entry "`camera_info`".
-
-
-
-</div><!-- file start -->
-
-<div id='pi_camera-decoder_node-autogenerated' markdown='1'>
-
-
-<!-- do not edit this file, autogenerated -->
-
-(Generated from [configuration `decoder_node.easy_node.yaml`](github:org=duckietown,repo=Software,path=decoder_node.easy_node.yaml,branch=master18).)
-
-A node that decodes a compressed image into a regular image. This is useful so that multiple nodes that need to use the image do not do redundant computaon.
-
-### Parameters {nonumber="1"}
-
-**Parameter `publish_freq`**: `float`; default value: `1.0`
-
-Frequency at which to publish (Hz).
-
-### Subscriptions {nonumber="1"}
-
-**Subscription `compressed_image`**: topic `~compressed_image` (`CompressedImage`)
-
-The image to decode.
-
-**Subscription `switch`**: topic `~switch` (`BoolStamped`)
-
-Switch to turn on or off. The node starts as active.
-
-### Publishers {nonumber="1"}
-
-**Publisher `raw`**: topic `~image/raw` (`Image`)
-
-The decoded image.
-
-
-
-</div><!-- file start -->
-
-<div id='pi_camera-camera_node_continuous-autogenerated' markdown='1'>
-
-
-<!-- do not edit this file, autogenerated -->
-
-(Generated from [configuration `camera_node_continuous.easy_node.yaml`](github:org=duckietown,repo=Software,path=camera_node_continuous.easy_node.yaml,branch=master18).)
-
-Camera driver.
-
-### Parameters {nonumber="1"}
-
-**Parameter `framerate`**: `float`; default value: `60.0`
-
-Frame rate
-
-**Parameter `res_w`**: `int`; default value: `320`
-
-Resolution (width)
-
-**Parameter `res_h`**: `int`; default value: `200`
-
-Resolution (height)
-
-### Subscriptions {nonumber="1"}
-
-No subscriptions defined.
-
-### Publishers {nonumber="1"}
-
-**Publisher `image_compressed`**: topic `~image/compressed"` (`CompressedImage`)
-
-TODO: Missing description for entry "`image_compressed`".
-
-
-
-</div><!-- file start -->
-
-<div id='pi_camera-img_process_node-autogenerated' markdown='1'>
-
-
-<!-- do not edit this file, autogenerated -->
-
-(Generated from [configuration `img_process_node.easy_node.yaml`](github:org=duckietown,repo=Software,path=img_process_node.easy_node.yaml,branch=master18).)
-
-Apparently, a template, or a node never finished.
-
-
-### Parameters {nonumber="1"}
-
-No parameters defined.
-
-### Subscriptions {nonumber="1"}
-
-No subscriptions defined.
-
-### Publishers {nonumber="1"}
-
-No publishers defined.
-
-
-
-</div>
