--- packages/pi_camera/README.md
@@ -1,27 +1,0 @@
-# Package `pi_camera` {#pi_camera}
-
-<move-here src='#pi_camera-autogenerated'/> 
-
-
-## Node `camera_node_sequence` {#pi_camera-camera_node_sequence}
-
-<move-here src="#pi_camera-camera_node_sequence-autogenerated"/>
-
-## Node `camera_node_continuous` {#pi_camera-camera_node_continuous}
-
-<move-here src="#pi_camera-camera_node_continuous-autogenerated"/>
-
-
-## Node `decoder_node` {#pi_camera-decoder_node}
-
-<move-here src="#pi_camera-decoder_node-autogenerated"/>
-
-
-## Node `img_process_node` {#pi_camera-img_process_node}
-
-<move-here src="#pi_camera-img_process_node-autogenerated"/>
-
-
-## Node `cam_info_reader_node` {#pi_camera-cam_info_reader_node}
-
-<move-here src="#pi_camera-cam_info_reader_node-autogenerated"/>
