--- packages/pi_camera/config/cam_info_reader_node/default.yaml
@@ -1,2 +1,0 @@
-#pi_camera/cam_info_reader_node
-pub_freq: 1.0
