--- packages/pi_camera/config/decoder_node/default.yaml
@@ -1,2 +1,0 @@
-#parameters for pi_camera/decoder_node
-publish_freq: 2.0
