--- packages/pi_camera/include/pi_camera/__init__.py
@@ -1,1 +1,0 @@
-from .camera_info import *
