--- packages/pi_camera/src/img_process_node.easy_node.yaml
@@ -1,6 +1,0 @@
-description: |
-    Apparently, a template, or a node never finished.
-parameters: {}
-subscriptions: {}
-publishers: {}
-contracts: {}
